test: remove failing test for "no response schema"
Issue
=

The test suite is failing both locally and in CI because a test file
`no-response-schema.yaml` seems to be missing from the repository.

The test seems to depend on this file, but the file is not included in
the repo. The snapshot, however, is included, so I'm thinking this must
have been some minor developer oversight.

Solution
=

Remove the failing test and its snapshot to make the build pass.